+++ CHANGES.rst
@@ -1,3 +1,11 @@
+19.1.0
+------
+
+* Add support for alternate persistence scopes for Windows
+  backend. Set ``.persist`` to "local machine" or "session"
+  to enable the alternate scopes or "enterprise" to use the
+  default scope.
+
 19.0.2
 ------